Remove redundant clamp from PID controller
The if statement below has the same effect, and it meant that the `else` branch
in the `if` below never executed. It appears that the design intention was that
the controller would not update the error when the control was clamped, and
that does indeed seem to help on longer runs.